moved bird classifier to a diff file
